feat: BodyType Entity, DTO 구현
